Added initial call to refresh, to pre-fix bug.
I wrote a menu module and discoverd a weird
bug in ncurses where updates to the virtual
screen before the very first call to refresh
are sometimes discarded, depending on the
combo of functions used...

Can't find any mention of this in the
documentation. But regardless, adding a
refresh() call before anything fixes the
problem, so I'm leaving it there.

Incidentally, I discarded the new module,
but the bug is likely to resurface.