Formatting characters in help are now pre-parsed
Lines in a Scroller buffer used to be parsed when the text
was displayed which caused long lines to be wrapped too
quickly, because it didn't take into account that formatting
characters aren't displayed. Now, scroller_write parses the
formatting characters out, storing them in an attribute
mask, and the wrapping works properly.