Scrolling help files is now less CPU-intensive
Scrolling with Page UP/DOWN no longer uses smooth scrolling,
which is very cpu intensive as it has to redraw the entire
window 5 times per scroll. I plan to add an option to
re-enable it though, as it looks much better.

Also, it now stores the length of lines, so it doesn't have
to call strlen all the time, which also contributed to CPU
saves.